+++ src/gql/types/generators/SubmissionGenerator.php
@@ -21,7 +21,7 @@ class SubmissionGenerator extends Generator implements GeneratorInterface, Singl
 
     public static function generateTypes(mixed $context = null): array
     {
-        $forms = Formie::$plugin->getForms()->getAllFormsWithLayouts();
+        $forms = Formie::$plugin->getForms()->getAllFormsWithLayouts(forSchema: true);
         $gqlTypes = [];
 
         foreach ($forms as $form) {